batt volt fix

+++ src/lora_interface.h
@@ -52,7 +52,7 @@ typedef struct {
     uint8_t pyro_state;        // 1 byte
     uint8_t sats;              // 1 byte
     uint8_t flight_state;      // 1 byte
-    float battery_voltage;     // 4 bytes
+    uint32_t battery_voltage;     // 4 bytes
 } goober_post_telemetry_payload_t; // total: 51 bytes
 
 typedef struct {
